# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_StickyFingersss_golang-todo-app_internal_core_transport_http_response.ErrorResponse:
    properties:
      error:
        example: full error text
        type: string
      message:
        example: short human readable message
        type: string
    type: object
...
  internal_features_tasks_transport_http.CreateTaskRequest:
    properties:
      author_user_id:
        example: 7
        type: integer
      completed:
        example: false
        type: boolean
      completed_at:
        example: "2026-08-21T15:30:00Z"
        type: string
      created_at:
        example: "2026-08-20T10:00:00Z"
        type: string
      description:
        example: Не забыть про скидку
        type: string
      id:
        example: 10
        type: integer
      title:
        example: Купить молоко
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_tasks_transport_http.CreateTaskResponse:
    properties:
      author_user_id:
        example: 7
        type: integer
      completed:
        example: false
        type: boolean
      completed_at:
        example: "2026-08-21T15:30:00Z"
        type: string
      created_at:
        example: "2026-08-20T10:00:00Z"
        type: string
      description:
        example: Не забыть про скидку
        type: string
      id:
        example: 10
        type: integer
      title:
        example: Купить молоко
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_tasks_transport_http.GetTaskResponse:
    properties:
      author_user_id:
        example: 7
        type: integer
      completed:
        example: false
        type: boolean
      completed_at:
        example: "2026-08-21T15:30:00Z"
        type: string
      created_at:
        example: "2026-08-20T10:00:00Z"
        type: string
      description:
        example: Не забыть про скидку
        type: string
      id:
        example: 10
        type: integer
      title:
        example: Купить молоко
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_tasks_transport_http.PatchTaskRequest:
    properties:
      completed:
        example: true
        type: boolean
      description:
        example: Не забыть про скидку
        type: string
      title:
        example: Купить молоко
        type: string
    type: object
  internal_features_tasks_transport_http.PatchTaskResponse:
    properties:
      author_user_id:
        example: 7
        type: integer
      completed:
        example: false
        type: boolean
      completed_at:
        example: "2026-08-21T15:30:00Z"
        type: string
      created_at:
        example: "2026-08-20T10:00:00Z"
        type: string
      description:
        example: Не забыть про скидку
        type: string
      id:
        example: 10
        type: integer
      title:
        example: Купить молоко
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_tasks_transport_http.TaskDTOResponse:
    properties:
      author_user_id:
        example: 7
        type: integer
      completed:
        example: false
        type: boolean
      completed_at:
        example: "2026-08-21T15:30:00Z"
        type: string
      created_at:
        example: "2026-08-20T10:00:00Z"
        type: string
      description:
        example: Не забыть про скидку
        type: string
      id:
        example: 10
        type: integer
      title:
        example: Купить молоко
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_users_transport_http_dto.CreateUserRequest:
    properties:
      full_name:
        example: Коля Колевич
        maxLength: 100
        minLength: 3
        type: string
      phone_number:
        example: "+77758451038"
        maxLength: 15
        minLength: 10
        type: string
    required:
    - full_name
    type: object
  internal_features_users_transport_http_dto.CreateUserResponse:
    properties:
      full_name:
        example: Колясик Колясикович
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+79132038908"
        type: string
      version:
        example: 3
        type: integer
    type: object
...
  internal_features_users_transport_http_dto.UserDTOResponse:
    properties:
      full_name:
        example: Колясик Колясикович
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+79132038908"
        type: string
      version:
        example: 3
        type: integer
    type: object
host: 127.0.0.1:5050
info:
  contact: {}
  description: Todo Application REST-API scheme
  title: Golang Todo API
  version: "1.0"
paths:
...
  /tasks:
    get:
      description: Просмотр списка задач с опциональной фильтрацией по автору и пагинацией
      parameters:
      - description: Фильтр по ID автора задачи
        in: query
        name: user_id
        type: integer
      - description: Размер страницы с задачами
        in: query
        name: limit
        type: integer
      - description: Смещение страницы с задачами
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное получение списка задач
          schema:
            items:
              $ref: '#/definitions/internal_features_tasks_transport_http.TaskDTOResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_StickyFingersss_golang-todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_StickyFingersss_golang-todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Получение задач
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Создать новую задачу в системе
      parameters:
      - description: CreateTask тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_tasks_transport_http.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Успешно созданная задача
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.CreateTaskResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_StickyFingersss_golang-todo-app_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Author user not found
          schema:
            $ref: '#/definitions/github_com_StickyFingersss_golang-todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_StickyFingersss_golang-todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Создать задачу
      tags:
      - tasks
  /tasks/{id}:
    delete:
      description: Удаление существующей в системе задачи по её ID
      parameters:
      - description: ID удаляемой задачи
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Успешное удаление задачи
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_StickyFingersss_golang-todo-app_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/github_com_StickyFingersss_golang-todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_StickyFingersss_golang-todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Удалить задачу
      tags:
      - tasks
    get:
      description: Получение задачи по её ID
      parameters:
      - description: ID получаемой задачи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Задача успешно найдена
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.GetTaskResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_StickyFingersss_golang-todo-app_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/github_com_StickyFingersss_golang-todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_StickyFingersss_golang-todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Получение задачи
      tags:
      - tasks
    patch:
      consumes:
      - application/json
      description: |-
        Изменение информации об уже существующей в системе задаче
        ### Логика обновления полей (Three-state logic):
        1. **Поле не передано**: значение в БД не меняется
        2. **Явно передано значение**: `"title": "Новое название"` — устанавливает новое значение
        3. **Передан null**: `"description": null` — очищает поле в БД (set to NULL)
        Ограничение: `title` и `completed` не могут быть выставлены как null
      parameters:
      - description: ID изменяемой задачи
        in: path
        name: id
        required: true
        type: integer
      - description: PatchTask тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_tasks_transport_http.PatchTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешно изменённая задача
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.PatchTaskResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_StickyFingersss_golang-todo-app_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/github_com_StickyFingersss_golang-todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_StickyFingersss_golang-todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Изменение задачи
      tags:
      - tasks
  /users:
    get:
      description: Просмотр списка пользователей с опциональной пагинацией
      parameters:
      - description: Размер страницы с пользователями
        in: query
        name: limit
        type: integer
      - description: Смещение страницы с пользователями
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное получение списка пользователей
          schema:
            items:
              $ref: '#/definitions/internal_features_users_transport_http_dto.UserDTOResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_StickyFingersss_golang-todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_StickyFingersss_golang-todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Получение пользователей
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Создать нового пользователя в системе
      parameters:
      - description: CreateUser тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_users_transport_http_dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Успешно созданный пользователь
          schema:
            $ref: '#/definitions/internal_features_users_transport_http_dto.CreateUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_StickyFingersss_golang-todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_StickyFingersss_golang-todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Создать пользователя
      tags:
      - users
...
swagger: "2.0"