# internal/features/tasks/transport/http/dto_common.go
package tasks_transport_http

import (
	"time"

	"github.com/StickyFingersss/golang-todo-app/internal/core/domain"
)

type TaskDTOResponse struct {
	ID           int        `json:"id" example:"10"`
	Version      int        `json:"version" example:"3"`
	Title        string     `json:"title" example:"Купить молоко"`
	Description  *string    `json:"description" example:"Не забыть про скидку"`
	Completed    bool       `json:"completed" example:"false"`
	CreatedAt    time.Time  `json:"created_at" example:"2026-08-20T10:00:00Z"`
	CompletedAt  *time.Time `json:"completed_at" example:"2026-08-21T15:30:00Z"`
	AuthorUserID int        `json:"author_user_id" example:"7"`
}

func taskDTOFromDomain(task domain.Task) TaskDTOResponse {
	return TaskDTOResponse{
		ID:           task.ID,
		Version:      task.Version,
		Title:        task.Title,
		Description:  task.Description,
		Completed:    task.Completed,
		CreatedAt:    task.CreatedAt,
		CompletedAt:  task.CompletedAt,
		AuthorUserID: task.AuthorUserID,
	}
}

func taskDTOsFromDomains(tasks []domain.Task) []TaskDTOResponse {
	dtos := make([]TaskDTOResponse, len(tasks))

	for i, task := range tasks {
		dtos[i] = taskDTOFromDomain(task)
	}

	return dtos
}